Refactor: 철자 틀린 컴포넌트, 파일명 수정
