+++ src/app/myPage/myCommentsPage/page.tsx
@@ -38,7 +38,7 @@ const MyCommentsPage = () => {
             <div>
               <img
                 src="https://images.unsplash.com/5/unsplash-kitsune-4.jpg?ixlib=rb-0.3.5&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=400&fit=max&ixid=eyJhcHBfaWQiOjEyMDd9&s=dd060fe209b4a56733a1dcc9b5aea53a"
-                alt="purfume image"
+                alt="perfume image"
               />
               <S.RatingWrap>
                 <p>{`${"브랜드명"}﹒${"브랜드명"}`}</p>
