+++ src/service/client/onBoarding.ts
@@ -5,13 +5,13 @@ import { perfumeBrands } from "@/types/res/perfume";
 import { setOnboardingRatingRequestDTO } from "@/types/req/onboarding";
 
 const endPoint = {
-  GET_PURFUME_BRANDS: "/perpicks/perfume-brands",
+  GET_PERFUME_BRANDS: "/perpicks/perfume-brands",
   SET_RATINGS_ONBOARDING: "/perpicks/ratings/onboarding",
 };
 
-async function getPurfumeBrands() {
+async function getPerfumeBrands() {
   const response = await axios.get<never, RestResponseType<perfumeBrands>>(
-    `${process.env.NEXT_PUBLIC_ENDPOINT_EXTERNAL}${endPoint.GET_PURFUME_BRANDS}`,
+    `${process.env.NEXT_PUBLIC_ENDPOINT_EXTERNAL}${endPoint.GET_PERFUME_BRANDS}`,
     {
       baseURL: process.env.NEXT_PUBLIC_ENDPOINT_EXTERNAL,
       headers: {
@@ -34,4 +34,4 @@ async function setOnboardingRating(payload: setOnboardingRatingRequestDTO) {
   return response;
 }
 
-export { getPurfumeBrands, setOnboardingRating };
+export { getPerfumeBrands, setOnboardingRating };
